+++ tests/conftest.py
@@ -85,7 +85,8 @@ def benchmark_config(auth, device) -> Generator[BenchmarkConfig, None, None]:
         evaluate_test_split=False,
         few_shot=True,
         num_iterations=1,
-        base_url=None,
+        api_base=None,
+        api_version=None,
         debug=False,
         run_with_cli=True,
     )